Made worktrees an Environment capability (orthogonal to Strategy) instead of a strategy (WIP)
