+++ cmd/juggler/extmanifest/extmanifest.go
@@ -32,6 +32,13 @@ type Provides struct {
 	ContextItems []string `json:"contextItems,omitempty"`
 	Strategies   []string `json:"strategies,omitempty"`
 	Commands     []string `json:"commands,omitempty"`
+	// Environments declares "where does this conversation's work physically
+	// happen" capabilities — git worktrees, devcontainers, remote/sandbox roots.
+	// This is a SEPARATE axis from Strategies (which govern loop autonomy:
+	// read-only / default / yolo): a conversation picks one environment AND one
+	// strategy independently, so "run in a worktree" composes with any autonomy
+	// level. Files end in `*-environment-type.js`.
+	Environments []string `json:"environments,omitempty"`
 	// SystemPrompt is a single module path (not a glob) whose default export
 	// `({enabledPluginIds}) => string` contributes terse, durable guidance to
 	// the system prompt — the extension's voice on how to use its tools. It is
@@ -81,6 +88,7 @@ func Validate(m Manifest, engineVersion string) error {
 	if len(m.Provides.ContextItems) == 0 &&
 		len(m.Provides.Strategies) == 0 &&
 		len(m.Provides.Commands) == 0 &&
+		len(m.Provides.Environments) == 0 &&
 		strings.TrimSpace(m.Provides.SystemPrompt) == "" {
 		return fmt.Errorf("manifest %q provides no capabilities", m.ID)
 	}
