+++ examples/extensions/juggler-worktrees/juggler.extension.json
@@ -8,6 +8,6 @@
   "engineApi": "^1.0.0",
   "permissions": ["shell.exec", "filesystem.write"],
   "provides": {
-    "strategies": ["strategies/*-strategy-type.js"]
+    "environments": ["environments/*-environment-type.js"]
   }
 }
