+++ web/engine.html
@@ -19,6 +19,7 @@
         "imports": {
             "juggler/context-item": "/v{{.StaticVersion}}/sdk/context-item.js",
             "juggler/strategy-type": "/v{{.StaticVersion}}/sdk/strategy-type.js",
+            "juggler/environment-type": "/v{{.StaticVersion}}/sdk/environment-type.js",
             "juggler/command-type": "/v{{.StaticVersion}}/sdk/command-type.js",
             "juggler/ops": "/v{{.StaticVersion}}/sdk/ops.js",
             "juggler/sandbox": "/v{{.StaticVersion}}/sdk/sandbox.js",
